requirements.txt
Pillow
PySide6
argparse